image repair func commited and funcs renamed

+++ Image_Stitching/Correlation.cpp
@@ -73,8 +73,8 @@ float Correlation(BYTE* img1, BYTE* img2, int width, int height, int zoneWidth, 
 
 	int newR1, newC1, newR2, newC2;
 	unsigned int total1 = 0, total2 = 0;
-	double total = 0.0F, totalSqr1 = 0.0F, totalSqr2 = 0.0F;
-	double mean1, mean2;
+	float total = 0.0F, totalSqr1 = 0.0F, totalSqr2 = 0.0F;
+	float mean1, mean2;
 
 #pragma omp parallel num_threads(NUM_THREADS) shared(zoneHeight,zoneWidth,start1H,start2H,start1W,start2W, width, img1, img2) private(newR1, newR2, newC1, newC2)
 	{
@@ -93,8 +93,8 @@ float Correlation(BYTE* img1, BYTE* img2, int width, int height, int zoneWidth, 
 		}
 	}
 
-	mean1 = double(total1 / (zoneHeight * zoneWidth));
-	mean2 = double(total2 / (zoneHeight * zoneWidth));
+	mean1 = float(total1 / (zoneHeight * zoneWidth));
+	mean2 = float(total2 / (zoneHeight * zoneWidth));
 
 #pragma omp parallel num_threads(NUM_THREADS) shared(zoneHeight,zoneWidth,start1H,start2H,start1W,start2W,width,img1,img2,mean1,mean2) private(newR1, newR2, newC1, newC2)
 	{
